rafactoring: mod tests extracted
